Update LocationFinder.tsx
Fetch requests work in such a way that when an option is clicked on the ComboBox, a request is fired. Next step is to make the request work in a way where when the input is changed, a new request is made. Main issue fixed: CORS error -- added proxy url to do so.