Fixed typo, functions, and switch cases
Fixed a typo in one of the stories. Changed indexIncrease and indexDecrease functions to use a local variable instead of currentIndex, and calls setCurrentIndex with the local variable at the end. Changed the border colour and accessibility switch case names to use the VendorApprovalStatus enum names instead of 'case 2/case 1'.